chore: advertise exports differently
