+++ frontend/src/scenes/events/EventsTable.tsx
@@ -2,7 +2,6 @@ import { useMemo } from 'react'
 import { useActions, useValues } from 'kea'
 import { EventDetails } from 'scenes/events/EventDetails'
 import { Link } from 'lib/components/Link'
-import { Popconfirm } from 'antd'
 import { FilterPropertyLink } from 'lib/components/FilterPropertyLink'
 import { Property } from 'lib/components/Property'
 import { autoCaptureEventToDescription } from 'lib/utils'
@@ -41,6 +40,7 @@ import { EventBufferNotice } from './EventBufferNotice'
 import { LemonDivider } from '@posthog/lemon-ui'
 import { sessionPlayerModalLogic } from 'scenes/session-recordings/player/modal/sessionPlayerModalLogic'
 import { SessionPlayerModal } from 'scenes/session-recordings/player/modal/SessionPlayerModal'
+import { ExportWithConfirmation } from '~/queries/nodes/DataTable/ExportWithConfirmation'
 
 export interface FixedFilters {
     action_id?: ActionType['id']
@@ -67,31 +67,6 @@ interface EventsTableProps {
     showPersonColumn?: boolean
     linkPropertiesToFilters?: boolean
     'data-attr'?: string
-}
-
-interface ExportWithConfirmationProps {
-    placement: 'topRight' | 'bottomRight'
-    onConfirm: (e?: React.MouseEvent<HTMLElement>) => void
-    children: React.ReactNode
-}
-
-function ExportWithConfirmation({ placement, onConfirm, children }: ExportWithConfirmationProps): JSX.Element {
-    return (
-        <Popconfirm
-            placement={placement}
-            title={
-                <>
-                    Exporting by csv is limited to 3,500 events.
-                    <br />
-                    To return more, please use <a href="https://posthog.com/docs/api/events">the API</a>. Do you want to
-                    export by CSV?
-                </>
-            }
-            onConfirm={onConfirm}
-        >
-            {children}
-        </Popconfirm>
-    )
 }
 
 export function EventsTable({
@@ -275,7 +250,7 @@ export function EventsTable({
     }, [tableWidth])
 
     const columns = useMemo(() => {
-        let columnsSoFar: LemonTableColumn<EventsTableRowItem, keyof EventsTableRowItem | undefined>[] = []
+        let columnsSoFar: LemonTableColumn<EventsTableRowItem, keyof EventsTableRowItem | undefined>[]
         if (selectedColumns === 'DEFAULT') {
             columnsSoFar = [...defaultColumns]
         } else {
@@ -529,7 +504,7 @@ export function EventsTable({
                                     defaultColumns={defaultColumns.map((e) => e.key || '')}
                                 />
                             )}
-                            {showExport ? (
+                            {showExport && (
                                 <LemonButtonWithPopup
                                     popup={{
                                         sameWidth: false,
@@ -541,6 +516,8 @@ export function EventsTable({
                                                 onConfirm={() => {
                                                     startDownload(exportColumns)
                                                 }}
+                                                actor={'events'}
+                                                limit={3500}
                                             >
                                                 <LemonButton fullWidth={true} status="stealth">
                                                     Export current columns
@@ -550,6 +527,8 @@ export function EventsTable({
                                                 key={0}
                                                 placement={'bottomRight'}
                                                 onConfirm={() => startDownload()}
+                                                actor={'events'}
+                                                limit={3500}
                                             >
                                                 <LemonButton fullWidth={true} status="stealth">
                                                     Export all columns
@@ -562,27 +541,6 @@ export function EventsTable({
                                 >
                                     Export
                                 </LemonButtonWithPopup>
-                            ) : (
-                                <Popconfirm
-                                    placement="topRight"
-                                    title={
-                                        <>
-                                            Exporting by csv is limited to 3,500 events.
-                                            <br />
-                                            To return more, please use{' '}
-                                            <a href="https://posthog.com/docs/api/events">the API</a>. Do you want to
-                                            export by CSV?
-                                        </>
-                                    }
-                                    onConfirm={() => startDownload()}
-                                >
-                                    <LemonButton
-                                        type="secondary"
-                                        icon={<IconExport style={{ color: 'var(--primary)' }} />}
-                                    >
-                                        Export
-                                    </LemonButton>
-                                </Popconfirm>
                             )}
                         </div>
                     </div>
